[add] detailのCRUDを削除し，Areaを複数持たせる (#68)
