+++ backend/domain/area.go
@@ -12,6 +12,8 @@ type Area struct {
 	City          string `json:"city" gorm:"not null"`
 	AddressNumber string `json:"address_number"`
 	BuildingName  string `json:"building_name"`
+	//belongs to
+	DetailId int `json:"detail_id" gorm:"foreign_key:ID"`
 	// //has one
 	// Detail Detail `json:"detail"`
 	//time stamp
